+++ include/build.h
@@ -2,6 +2,6 @@
  * Author: sascha_lammers@gmx.de
  */
 #pragma once
-#define __BUILD_NUMBER "11398"
-#define __BUILD_NUMBER_INT 11398
+#define __BUILD_NUMBER "11425"
+#define __BUILD_NUMBER_INT 11425